backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
pydantic==2.10.6
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1 